% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/then.R
\name{compose_visitor}
\alias{\%then\%}
\alias{compose_visitor}
\title{compose visitor functions}
\usage{
lhs \%then\% rhs
}
\arguments{
\item{lhs}{visitor object}

\item{rhs}{visitor object}
}
\value{
composed visitor object
}
\description{
compose visitor functions
}
\examples{
add1 <- make_visitor(
  leaf = function(x) if(is.numeric(x)) x + 1 else x
)
mul2 <- make_visitor(
  leaf = function(x) if(is.numeric(x)) x * 2 else x
)
walk_ast(quote(1 + 2 * 3), add1)
walk_ast(quote(1 + 2 * 3), mul2)
walk_ast(quote(1 + 2 * 3), add1 \%then\% mul2)
walk_ast(quote(1 + 2 * 3), mul2 \%then\% add1)
}

