% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{helper}
\alias{to_call}
\alias{to_list}
\title{helper functions for walkast}
\usage{
to_list()

to_call()
}
\description{
helper functions for walkast
}
\examples{
walk_ast(quote(1 + 2 * 3), to_list())
walk_ast(list(quote(`+`), 1, list(quote(`*`), 2, 3)), to_call())
walk_ast(quote(1 + 2 * 3), to_list() \%then\% to_call())
}

