% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{nest_expr}
\alias{replace}
\alias{replace_post}
\alias{replace_pre}
\title{replace call object
Replacement}
\usage{
replace_pre(before, after)

replace_post(before, after)

nest_expr(before, after, count)
}
\arguments{
\item{before}{language object or literal}

\item{after}{language object or literal}

\item{count}{count of iteration}
}
\description{
replace call object
Replacement
}
\examples{
walk_ast(quote(x + y + z), replace_pre(quote(y), quote(www)))
walk_ast(quote(x + x + x), replace_pre(quote(x + x), quote(x)))
walk_ast(quote(x + x + x), replace_post(quote(x + x), quote(x)))

walk_ast(quote(1 + 2 + x^3), replace_pre(quote(x), quote(y)))
walk_ast(quote(1 + 2 + x^3), replace_pre(quote(x^3), quote(y)))
walk_ast(quote(1 + 2 + x^3), replace_pre(quote(2), quote(99)))
walk_ast(quote(1 + 2 + x^3), replace_pre(quote(1+2), quote(99)))
walk_ast(quote(1 + 2 + x^3), replace_pre(quote(`+`), quote(`-`)))

nest_expr(quote((1 + x)^2), quote(x), 3)
nest_expr(quote((1 + x)^2), quote(1 + x), 3)
eval(nest_expr(quote(1 + 1 / x), quote(x), 40), list(x = 1)) == (1 + sqrt(5)) / 2
}

